app/requirements.txt
click==7.1.2
contextlib2==0.6.0.post1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyYAML==5.4
schema==0.7.2
Werkzeug==1.0.1
